add test to create task on post.
