+++ views.py
@@ -1,7 +1,3 @@
-from scheduler_wrapper import scheduler
-import datetime
-
-
 from flask import Blueprint, request
 from log.log_handler import log
 import request_handler
@@ -15,8 +11,6 @@ views_blueprint = Blueprint("views", __name__)
 @views_blueprint.route("/")
 @log
 def hello():
-    run_time = datetime.datetime.now() + datetime.timedelta(seconds=5)
-    scheduler.add_job(print_hello_world, "date", run_date=run_time)
     return "Hello, World!"
 
 
@@ -24,7 +18,3 @@ def hello():
 @log
 def task():
     return request_handler.post_task(request)
-
-
-def print_hello_world():
-    print("Hello, World!")
